# reports/openapi_day4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCM Copilot",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/ops/chat": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "业务对话（SSE 流式）",
        "description": "意图识别 → 审批门 → 执行 → 回答。高危操作在 approval_gate 中断并推送 approval_request 事件（HITL），前端展示审批表单后调 POST /api/v1/ops/approval 恢复。事件协议：progress / approval_request / message / done / error。需要权限 ops:tool:execute。",
        "operationId": "chat_api_v1_ops_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE 流（text/event-stream）：\n- progress:         {type, node, data:{result}}——intent/approval_gate/execute/respond 节点\n- approval_request: {type, approval_id, form, session_id}——HITL 审批表单\n- message:          {type, role, content, delta, session_id}——打字机增量\n- done / error:     流结束 / 链路异常",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "data: {\"type\":\"progress\",\"node\":\"intent\",\"data\":{\"result\":\"识别意图：改单\"}}\n\n"
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ops/approval": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "审批决策（HITL 恢复）",
        "description": "批准/拒绝高危操作，Command(resume) 恢复 LangGraph 图继续执行。需要权限 ops:approval:manage。",
        "operationId": "approval_action_api_v1_ops_approval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ops/report": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "异步报表（入队削峰）",
        "description": "立即返回 task_id，worker 后台生成；队列不可用 → 同步降级返回 result。需要权限 ops:tool:execute。",
        "operationId": "report_async_api_v1_ops_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportEnqueueOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ops/report/sync": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "同步报表",
        "description": "直接生成返回（不削峰，对比/调试用）。需要权限 ops:tool:execute。",
        "operationId": "report_sync_api_v1_ops_report_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSyncOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ops/report/{task_id}": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "轮询异步报表结果",
        "description": "finished → result（展开进顶层）；未完成 → {ready: false, status}。",
        "operationId": "report_status_api_v1_ops_report__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStatusOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalIn": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "HITL 会话标识"
          },
          "approval_id": {
            "type": "string",
            "title": "Approval Id",
            "description": "审批单 ID（approval_request 事件下发）"
          },
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "description": "批准 / 拒绝"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "审批意见",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "approval_id",
          "decision"
        ],
        "title": "ApprovalIn",
        "description": "HITL 审批决策请求体。"
      },
      "ApprovalOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "reply": {
            "type": "string",
            "title": "Reply",
            "description": "执行后的自然语言回复",
            "default": ""
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "default": false
          },
          "tool_result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Tool Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ApprovalOut",
        "description": "审批决策响应（resume LangGraph 图后的执行结果）。"
      },
...
      "Err": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "机器可读错误码，如 AUTH_401 / QUOTA_429"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "人类可读错误说明"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "贯穿一次请求的 request_id（排查定位用）"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "Err",
        "description": "统一错误响应体（OpenAPI `responses={401: {\"model\": Err}, ...}` 引用）。"
      },
...
      "OpsChatIn": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "业务指令（查单/改单/报表等）"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话标识；缺省服务端生成"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OpsChatIn",
        "description": "业务对话请求体（SSE 流式）。"
      },
...
      "ReportEnqueueOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "async": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Async",
            "description": "是否异步入队"
          },
          "sync": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ReportEnqueueOut",
        "description": "异步报表入队响应（队列不可用时同步降级返回 result）。"
      },
      "ReportIn": {
        "properties": {
          "report_type": {
            "type": "string",
            "enum": [
              "inventory",
              "reconciliation"
            ],
            "title": "Report Type",
            "default": "inventory"
          },
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From",
            "description": "起始日期 YYYY-MM-DD"
          },
          "to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To",
            "description": "截止日期 YYYY-MM-DD"
          }
        },
        "type": "object",
        "title": "ReportIn",
        "description": "报表请求体（async / sync 共用）。"
      },
      "ReportStatusOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "ready": {
            "type": "boolean",
            "title": "Ready",
            "default": false
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ReportStatusOut",
        "description": "异步报表轮询响应（finished 时展开 result 字段，extra 放行）。"
      },
      "ReportSyncOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ReportSyncOut",
        "description": "同步报表响应。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}